requirements.txt
python-telegram-bot[job-queue]>=21.3
anthropic>=0.49.0
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
PyMuPDF>=1.24.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
flask>=3.0.0
